requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
chatgpt_md_converter==0.1.2
click==8.1.7
duckduckgo_search==6.2.6
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pillow==10.4.0
primp==0.5.5
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.5
tinydb==4.8.0
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
